Moved the fun/global vars to orig location and removed tests
