+++ portsyncd/linksync.h
@@ -4,16 +4,8 @@
 #include "dbconnector.h"
 #include "producerstatetable.h"
 #include "netmsg.h"
-#include "exec.h"
-#include "warm_restart.h"
-#include "shellcmd.h"
 
 #include <map>
-#include <set>
-#include <vector>
-#include <set>
-#include <map>
-#include <list>
 
 namespace swss {
 
@@ -36,6 +28,6 @@ private:
 
 }
 
-void handlePortConfigFromConfigDB(swss::ProducerStateTable &, swss::DBConnector &, bool );
+
 
 #endif
